refactor: some cleanup after AI
