+++ pkg/cloudprovider/suite_test.go
@@ -171,3 +171,12 @@ func validateVMNodeClaim(nodeClaim *karpv1.NodeClaim, nodePool *karpv1.NodePool)
 	// VM-specific validation (should NOT have AKS machine annotation)
 	Expect(nodeClaim.Annotations).ToNot(HaveKey(v1beta1.AnnotationAKSMachineResourceID))
 }
+
+func validateAKSMachineNodeClaim(nodeClaim *karpv1.NodeClaim, nodePool *karpv1.NodePool) {
+	// Common validations
+	validateNodeClaimCommon(nodeClaim, nodePool)
+
+	// AKS-specific annotations
+	Expect(nodeClaim.Annotations).To(HaveKey(v1beta1.AnnotationAKSMachineResourceID))
+	Expect(nodeClaim.Annotations[v1beta1.AnnotationAKSMachineResourceID]).ToNot(BeEmpty())
+}
